+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - [TD-4176]
   - Added descrition in data structure tag
   - Changed description for comment in data structures tags link
+- [TD-4661] Support for ReferenceDatasets on Implementations
 
 ### Fixed